replaced with my own modifications

+++ GCoptimization.h
@@ -163,11 +163,11 @@ public:
 	typedef GCO_ENERGYTERMTYPE EnergyTermType;
 #else
 #ifdef GCO_ENERGYTYPE32
-	typedef int EnergyType;        // 32-bit energy total
+	typedef float EnergyType;        // 32-bit energy total
 #else
-	typedef long long EnergyType;  // 64-bit energy total
+	typedef float EnergyType;  		// 64-bit energy total
 #endif
-	typedef int EnergyTermType;    // 32-bit energy terms
+	typedef float EnergyTermType;    // 32-bit energy terms
 #endif
 	typedef Energy<EnergyTermType,EnergyTermType,EnergyType> EnergyT;
 	typedef EnergyT::Var VarID;
